% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpsmar90}
\alias{cpsmar90}
\title{CPS March Supplement (ASEC) from NBER, 1990}
\format{A data frame}
\source{
\url{http://www.nber.org/data/current-population-survey-data.html}
}
\usage{
cpsmar90
}
\description{
\code{cpsmar90} contains all the variables (as of Aug. 30, 2015)
from NBER-CPS ASEC, 1990
}
\keyword{datasets}

