% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpsmar92}
\alias{cpsmar92}
\title{CPS March Supplement (ASEC) from NBER, 1992}
\format{A data frame}
\source{
\url{http://www.nber.org/data/current-population-survey-data.html}
}
\usage{
cpsmar92
}
\description{
\code{cpsmar92} contains all the variables (as of Aug. 30, 2015)
from NBER-CPS ASEC, 1992
}
\keyword{datasets}

