% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpsmar96}
\alias{cpsmar96}
\title{CPS March Supplement (ASEC) from NBER, 1996}
\format{A data frame}
\source{
\url{http://www.nber.org/data/current-population-survey-data.html}
}
\usage{
cpsmar96
}
\description{
\code{cpsmar96} contains all the variables (as of Aug. 30, 2015)
from NBER-CPS ASEC, 1996
}
\keyword{datasets}

