% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpsmar99}
\alias{cpsmar99}
\title{CPS March Supplement (ASEC) from NBER, 1999}
\format{A data frame}
\source{
\url{http://www.nber.org/data/current-population-survey-data.html}
}
\usage{
cpsmar99
}
\description{
\code{cpsmar99} contains all the variables (as of Aug. 30, 2015)
from NBER-CPS ASEC, 1999
}
\keyword{datasets}

